[IMP] contract: Add templating * Add template functionality for contracts
