+++ contract/README.rst
@@ -66,6 +66,7 @@ Contributors
 * Pedro M. Baeza <pedro.baeza@tecnativa.com>
 * Carlos Dauden <carlos.dauden@tecnativa.com>
 * Angel Moya <angel.moya@domatix.com>
+* Dave Lasley <dave@laslabs.com>
 
 Maintainer
 ----------
